pyproject.toml
[tool.poetry]
name = "llm_classification"
version = "0.1.0"
description = "LLM-based grievance classification service"
authors = ["Anirudh-Murali <anirudhvmurali@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
aiohttp = "^3.9.0"
tqdm = "^4.66.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
